Move `cmd + .` shortcut handler to `MainNavbar`
